initial

+++ Movies-Reveiws-Web/src/main/java/com/Movies/Reveiw/Web/Movies/Reveiws/Web/repositories/ReviewRepository.java
@@ -7,5 +7,5 @@ import com.Movies.Reveiw.Web.Movies.Reveiws.Web.models.Movie;;
 import java.util.List;
 
 public interface ReviewRepository extends JpaRepository<Review, Long > {
-    List<Review> findByMovieId(Long movieId);
+    List<Review> findByMovieId(int movieId);
 }
